requirements.txt
Flask-RESTful==0.3.8
Flask==1.1.2
mysql-connector-python==8.0.22
mysqlclient==2.0.3

